fix: CommandLineApplication을 SpringBoot를 실행하는 Application과 명령어를 입력받고 이에 맞는 service logic을 수행하는 CommandLineController로 나누다. [Week1.Day4]
