+++ kdt-spring-order/src/main/java/org/prgrms/kdt/Application.java
@@ -1,0 +1,14 @@
+package org.prgrms.kdt;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Application {
+
+    public static void main(String[] args) {
+        var springApplication = new SpringApplication(Application.class);
+        springApplication.setAdditionalProfiles("dev");
+        springApplication.run(args);
+    }
+}
